src/python/requirements.txt

pandas>=1.3.0
numpy>=1.20.0
requests>=2.25.0
beautifulsoup4>=4.9.0
pillow>=8.0.0
torch>=1.9.0
torchvision>=0.10.0
